Add data classes for payment processing and refactor related components
