+++ src/Data/PaymentAttributesData.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Bnzo\Fintecture\DTO;
+namespace Bnzo\Fintecture\Data;
 
 use Spatie\LaravelData\Data;
 use Spatie\LaravelData\Optional;
 
-class PaymentAttributesDTO extends Data
+class PaymentAttributesData extends Data
 {
     public function __construct(
         public string $amount,
