+++ src/Data/PaymentData.php
@@ -1,21 +1,21 @@
 <?php
 
-namespace Bnzo\Fintecture\DTO;
+namespace Bnzo\Fintecture\Data;
 
 use Spatie\LaravelData\Attributes\MapInputName;
 use Spatie\LaravelData\Data;
 use Spatie\LaravelData\Support\Transformation\TransformationContext;
 use Spatie\LaravelData\Support\Transformation\TransformationContextFactory;
 
-class PaymentDTO extends Data
+class PaymentData extends Data
 {
     public function __construct(
         #[MapInputName('meta')]
-        public PaymentSettingsDTO $settings,
+        public PaymentSettingsData $settings,
         #[MapInputName('data.attributes')]
-        public PaymentAttributesDTO $attributes,
+        public PaymentAttributesData $attributes,
         #[MapInputName('meta')]
-        public PaymentCustomerDTO $customer) {}
+        public PaymentCustomerData $customer) {}
 
     public function transform(
         null|TransformationContextFactory|TransformationContext $transformationContext = null,
